requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.6
pillow>=10.0.0
opencv-python>=4.8.0
easyocr>=1.7.0
pytesseract>=0.3.10
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0
numpy>=1.24.0
aiofiles>=23.1.0
jinja2>=3.1.0
python-jose>=3.3.0
psutil>=5.9.0